% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insertDoubleHash}
\alias{insertDoubleHash}
\title{Add a double hash comment}
\usage{
insertDoubleHash()
}
\description{
The function inserts or removes two hashes and a space (\code{"## "}) at the
beginning of a line. It also works for multiple line. The function is meant
to behave similarly to the Rstudio built-in \code{Comment / Uncomment
Selection} tool (see Details for an comprehensive description).
}
\details{
The function has two behaviors depending on the current selection in the
document:
\itemize{
  \item The current selection spans a single line. Different scenarios
  \itemize{
    \item The line doesn't start with \code{"#"} (ignoring leading spaces):
    the \code{"## "} is inserted after the leading space
    \item The line starts with \code{"## "} (ignoring leading spaces):
    the leading \code{"## "} is removed
    \item The line start with a regular comment \code{"#"}  (ignoring leading
    spaces): the comment (\code{"# "} or \code{"#"}) is replaced by
    \code{"## "}. For roxygen2 comments \code{"#'"} nothing happen.
  }
  \item The current selection spans multiple lines:
  \itemize{
    \item Any of the lines don't start with \code{"#"} (ignoring leading
    spaces): the \code{"## "} is inserted after the smallest leading space
    for all lines. For roxygen2 comment \code{"#'"} nothing happen.
    \item All lines start with \code{"## "} (ignoring leading spaces):
    all leading \code{"## "} are removed.
  }
}
}
